requirements.txt
pyusb>=1.2.1
python-can>=4.3.0
